move LTRQParserPlugin to solr.search (and implementing classes to solr.search.ltr) under solr/contrib
