+++ solr/contrib/ltr/example/solrconfig.xml
@@ -857,7 +857,7 @@
     </requestHandler>
 
  <!-- Query parser used to rerank top docs with a provided model -->
-  <queryParser name="ltr" class="org.apache.solr.ltr.ranking.LTRQParserPlugin" >
+  <queryParser name="ltr" class="org.apache.solr.search.LTRQParserPlugin" >
   <int name="LTRMaxThreads">10</int> <!-- Maximum threads to use for all queries -->
   <int name="LTRMaxQueryThreads">10</int> <!-- Maximum threads to use for a single query-->
   </queryParser>
