+++ solr/contrib/ltr/src/java/org/apache/solr/search/ltr/LTRCollector.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.solr.ltr.ranking;
+package org.apache.solr.search.ltr;
 
 import java.io.IOException;
 import java.lang.invoke.MethodHandles;
